small edit to avoid off-by-one error on final loop
the code finished running, save for the very last name in the input
file. It looks like there was an off-by-one error with the bounds of
df_chunk, giving it a length of 0, which then caused an error due to
dividing by zero when we tried to calculate the time it took to
process then names it contained. Hopefully should be fixed now.